Refactor Cloudinary integration in communityEventController and itemsController
- Moved Cloudinary URL parsing to a shared utility function for better code reuse.
- Updated image and video upload functions to use resolved folder paths from environment variables.
- Enhanced folder handling for community banners and videos to ensure consistent uploads.
